Add support for executing cmd with vars inlined
Previously exec would consider the first flag arg to be the command and
all later args to be command args.

This commit updates that behavior so that individual environment
variables can be overwritten as needed. For example, an envy namespace
may have a baseline value for the variable FOO that should be
overwritten with an ephemeral value that shouldn't be stored.

There is support for the following formats:

- env FOO=BAR [command] [args...]
- FOO=BAR [command] [args...]